Implements changes to ADR-0008 and ADR-M-0002
