+++ tests/README.md
@@ -96,7 +96,7 @@ conda run -n dev pytest -q -m "notebook"
 | `test_engine_contracts.py::test_engine_provider_diagnostics_shape` | X |  |  |  |  |
 | `test_engine_contracts.py::test_engine_base_is_abstract` | X |  |  |  |  |
 | `test_engine_contracts.py::test_engine_base_contract_with_dummy_subclass` | X |  |  |  |  |
-| `test_engine_contracts.py::test_engine_ort_run_tile_contract` | X |  |  |  |  |
+| `test_model_resunet.py::test_engine_ort_run_tile_contract` | X |  |  |  |  |
 | `test_hrdem_mosaic.py::test_build_fetch_tile_grid_gdf_and_selection_mask_writes_geojson` | X |  |  |  | X |
 | `test_hrdem_mosaic.py::test_download_hrdem_project_extent_for_data_case` | X |  | X |  | X |
 | `test_hrdem_mosaic.py::test_fetch_hrdem_synthetic_cases` |  |  | X |  |  |
@@ -118,7 +118,9 @@ conda run -n dev pytest -q -m "notebook"
 | `test_preprocessing.py::test_write_platform_prepared_rasters_honors_crs_policy` | X |  |  |  |  |
 | `test_preprocessing.py::test_write_dem_from_asset_hrefs_outputs_float32_non_empty` | X |  |  |  |  |
 | `test_tohr_regression.py::test_tohr_regression_matches_case_spec_metrics` |  |  | X |  | X |
-| `test_tohr_regression.py::test_tohr_on_the_fly_synthetic_tiles` |  |  | X |  |  |
+| `test_model_resunet.py::test_resunet_tohr_on_the_fly_synthetic_tiles` |  |  | X |  |  |
+| `test_model_resunet.py::test_resunet_tohr_hard_windowed_tiles` |  |  | X |  |  |
+| `test_model_costgrow.py::test_costgrow_tohr_uses_windowed_path_for_large_hard_rasters` | X |  |  |  |  |
 
 
 # simple container tests
