+++ tests/test_tohr_regression.py
@@ -1,8 +1,6 @@
-"""Tests for ToHR regression and synthetic tiling behavior."""
+"""Tests for ToHR regression against committed case specs."""
 
 import importlib.util
-import json
-import logging
 import os
 import subprocess
 import sys
@@ -13,10 +11,9 @@ np = pytest.importorskip("numpy")
 rasterio = pytest.importorskip("rasterio")
 
 import floodsr.dem_sources.catalog
-import floodsr.models.ResUNet_16x_DEM as resunet_module
 import floodsr.tohr
 import misc.eval
-from conftest import default_model_version, logger, synthetic_tohr_tiles, synthetic_tohr_windowed_tiles, tile_case_d, tohr_model_fp
+from conftest import logger, tile_case_d
 from floodsr.model_registry import model_version_requires_artifact
 
 pytestmark = pytest.mark.network
@@ -187,69 +184,5 @@ def test_tohr_regression_matches_case_spec_metrics(
     assert rounded_actual == rounded_expected
 
 
-@pytest.mark.parametrize(
-    "window_method, tile_overlap, expected_execution_path",
-    [
-        pytest.param("hard", 0, "simple", id="on_the_fly_synth_hard"),
-        pytest.param("feather", 1, "simple", id="on_the_fly_synth_feather"),
-    ],
-)
-def test_tohr_on_the_fly_synthetic_tiles(
-    tohr_model_fp: Path,
-    default_model_version: str,
-    synthetic_tohr_tiles: dict,
-    window_method: str,
-    tile_overlap: int,
-    expected_execution_path: str,
-    logger,
-) -> None:
-    """Run tiled ToHR on on-the-fly synthetic rasters for both window methods."""
-    # `tohr_model_fp` is provided by tests/conftest.py::tohr_model_fp.
-    # Execute end-to-end inference for the selected synthetic tiling setup.
-    result = floodsr.tohr.tohr(
-        model_version=default_model_version,
-        model_fp=tohr_model_fp,
-        depth_lr_fp=synthetic_tohr_tiles["depth_lr_fp"],
-        dem_hr_fp=synthetic_tohr_tiles["dem_fp"],
-        output_fp=synthetic_tohr_tiles["output_fp"],
-        window_method=window_method,
-        tile_overlap=tile_overlap,
-        logger=logger,
-    )
-
-    # Validate basic output contract for the synthetic test fixture.
-    with rasterio.open(result["output_fp"]) as ds:
-        pred = ds.read(1)
-    assert pred.shape == synthetic_tohr_tiles["hr_shape"]
-    assert pred.dtype == np.float32
-    assert pred.size > 0
-    assert result["execution_path"] == expected_execution_path
 
 
-def test_tohr_hard_windowed_tiles(
-    tohr_model_fp: Path,
-    default_model_version: str,
-    synthetic_tohr_windowed_tiles: dict,
-    logger,
-):
-    """Run hard-window ToHR on a synthetic case that should trigger windowed IO."""
-    result = floodsr.tohr.tohr(
-        model_version=default_model_version,
-        model_fp=tohr_model_fp,
-        depth_lr_fp=synthetic_tohr_windowed_tiles["depth_lr_fp"],
-        dem_hr_fp=synthetic_tohr_windowed_tiles["dem_fp"],
-        output_fp=synthetic_tohr_windowed_tiles["output_fp"],
-        window_method="hard",
-        tile_overlap=0,
-        logger=logger,
-    )
-
-    with rasterio.open(result["output_fp"]) as ds:
-        pred = ds.read(1)
-    assert pred.dtype == np.float32
-    assert pred.size > 0
-    assert result["execution_path"] == "windowed"
-    assert Path(result["output_fp"]).suffix == (".vrt" if resunet_module.gdal is not None else ".tif")
-
-
-
